+++ CHANGELOG.md
@@ -8,7 +8,22 @@
 
 ### Fixes
 
-### Depracated
+### Deprecated
+
+## v0.3.0 - 2021-11-25
+
+### New features
+
+* Added `onlyDisplayIn` and `onlyDisplayOn` to the schema and updated the `getContent` method on document and publications to allow showing flashnews only in designated language or for a particular language. These are arrays to allow for multiple languages to be specified.
+* Added `defaultLanguage` to `freedombase:flashnews-create` that existed previously, but now supplements `onlyDisplayIn` when the display language does not equals any of the defined languages by showing content in the default language.
+* Added some initial tests
+
+## v0.2.1 - 2021-11-24
+
+### Fixes
+
+* Sorting of news according to date start
+
 
 ## v0.2.0 - 2021-11-23